Use uWS for Node.js' websocket benchmark
Currently, the websocket benchmark for Node.js uses `ws`, which seems to
be a slower library than `uWS` which is what Bun uses. There is no
reason in my opinion why `ws` is to be used over `uWS` as Node.js does
not officialy endorse either and both are community packages. The only
difference is that `ws` is more popular.

Here are the results when Node.js' benchmark is switched to `uWS`:

```
node (ws): [80841, 95556, 88358, 88848, 103426, 80422, 92639, 96288, 85550, 95432]
node (uws):  [409600, 819200, 819200, 819200, 819200, 819200, 819200, 819200, 819200, 819200]
```

As you can see, there is a significant jump by using `uWS` on Node.js.

Do note that there seems to be a limit (`819200`) that's being hit on
my machine that I cannot figure out. I tried changing my ulimits to
infinity to no avail. However, these benchmarks should be enough of a
proof of concept to prove that using `uWS` does seem to bring about a
significant improvement in perf.